% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMDV.R
\name{loadMDV}
\alias{loadMDV}
\title{Loads some Magnus D Vigeland (MDV) libraries from GitHub}
\usage{
loadMDV(Force = FALSE, github = TRUE, build = TRUE)
}
\arguments{
\item{Force}{logical}

\item{github}{logical}

\item{build}{logical}
}
\value{
Finished
}
\description{
Loads pedsuite libraries and some more
}
\examples{
 library(devtools)
loadMDV()

}
